Flat Website - Version 3: Draggable Wall
In this version the screens are positioned inside a parent cylinder wall. Grab interactions were removed from each screen, only the parent cylinder is grabbable and preserves its initial orientation in space. Scale of each screen is set to 1 - to check if quality will be better than in the previous version.